+++ aries-site/src/examples/components/layer/components/LayerHeader.js
@@ -36,5 +36,5 @@ LayerHeader.propTypes = {
   informational: PropTypes.bool,
   onClose: PropTypes.func,
   title: PropTypes.string,
-  subtitle: PropTypes.string,
+  subtitle: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),
 };
