+++ aries-site/src/pages/templates/user-feedback-collection.mdx
@@ -1,4 +1,4 @@
-import { PowerOnExample, SimpleFeedbackExample, SubscriptionOrderExample, UnSolicitedFeedbackExample } from '../../examples';
+import { SimpleFeedbackExample, SolicitedFeedbackExample } from '../../examples';
 import { Example } from '../../layouts';
 
 <Example
@@ -48,7 +48,7 @@ Click on "Power 1 Device" to see the progressive example of the feedback questio
   height="650px"
   width="100%"
 >
-  <PowerOnExample />
+  <SolicitedFeedbackExample />
 </Example>
